Restructured main logic of running interpreter
